requirements.txt
requests==2.31.0
pandas==2.2.1
pydantic==2.6.4
python-dotenv==1.0.1
pyarrow==15.0.2
seaborn==0.13.2
matplotlib==3.10.9
duckdb==1.5.2
ipykernel==6.29.5
nbformat==5.10.4